python/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
numpy
scikit-learn
PyPDF2
fastembed
